requirements.txt
dash
plotly.express
pandas
numpy
matplotlib.pyplot
seaborn
